fix: types and asserts, passing UT
