+++ src/utils/common.ts
@@ -13,8 +13,9 @@ import { isNode } from 'std-env';
 import { getRandomBytesBrowser } from './browser';
 import { SWARM_ZERO_ADDRESS } from './constants';
 import { getRandomBytesNode } from './node';
-import { FeedPayloadResult, FileInfo, ShareItem, WrappedFileInfoFeed, WrappedUploadResult } from './types';
+import { FeedPayloadResult, WrappedUploadResult } from './types';
 import { FileInfoError } from './errors';
+import { asserWrappedUploadResult } from './asserts';
 
 // Fetches the feed data for the given topic, index and address
 export async function getFeedData(
@@ -49,93 +50,6 @@ export function generateTopic(): Topic {
   return new Topic(getRandomBytesBrowser(Topic.LENGTH));
 }
 
-export function isObject(value: unknown): value is Record<string, unknown> {
-  return value !== null && typeof value === 'object';
-}
-
-export function isStrictlyObject(value: unknown): value is Record<string, unknown> {
-  return isObject(value) && !Array.isArray(value);
-}
-
-export function isRecord(value: unknown): value is Record<string, string> {
-  return isStrictlyObject(value) && Object.values(value).every((v) => typeof v === 'string');
-}
-
-export function assertFileInfo(value: unknown): asserts value is FileInfo {
-  if (!isStrictlyObject(value)) {
-    throw new TypeError('FileInfo has to be object!');
-  }
-
-  const fi = value as unknown as FileInfo;
-
-  new Reference(fi.file.reference);
-  new Reference(fi.batchId);
-  new Reference(fi.file.historyRef);
-
-  if (fi.topic !== undefined) {
-    new Topic(fi.topic);
-  }
-
-  if (fi.customMetadata !== undefined && !isRecord(fi.customMetadata)) {
-    throw new TypeError('FileInfo customMetadata has to be object!');
-  }
-
-  if (fi.timestamp !== undefined && typeof fi.timestamp !== 'number') {
-    throw new TypeError('timestamp property of FileInfo has to be number!');
-  }
-
-  if (fi.owner !== undefined) {
-    new EthAddress(fi.owner);
-  }
-
-  if (fi.name !== undefined && typeof fi.name !== 'string') {
-    throw new TypeError('fileName property of FileInfo has to be string!');
-  }
-
-  if (fi.preview !== undefined) {
-    new Reference(fi.preview.reference);
-    new Reference(fi.preview.historyRef);
-  }
-
-  if (fi.shared !== undefined && typeof fi.shared !== 'boolean') {
-    throw new TypeError('shared property of FileInfo has to be boolean!');
-  }
-
-  if (fi.redundancyLevel !== undefined && typeof fi.redundancyLevel !== 'number') {
-    throw new TypeError('redundancyLevel property of FileInfo has to be number!');
-  }
-}
-
-export function assertShareItem(value: unknown): asserts value is ShareItem {
-  if (!isStrictlyObject(value)) {
-    throw new TypeError('ShareItem has to be object!');
-  }
-
-  const item = value as unknown as ShareItem;
-
-  assertFileInfo(item.fileInfo);
-
-  if (item.timestamp !== undefined && typeof item.timestamp !== 'number') {
-    throw new TypeError('timestamp property of ShareItem has to be number!');
-  }
-
-  if (item.message !== undefined && typeof item.message !== 'string') {
-    throw new TypeError('message property of ShareItem has to be string!');
-  }
-}
-
-export function assertWrappedFileInoFeed(value: unknown): asserts value is WrappedFileInfoFeed {
-  if (!isStrictlyObject(value)) {
-    throw new TypeError('WrappedMantarayFeed has to be object!');
-  }
-
-  const wmf = value as unknown as WrappedFileInfoFeed;
-
-  if (wmf.eGranteeRef !== undefined) {
-    new Reference(wmf.eGranteeRef);
-  }
-}
-
 // status is undefined in the error object
 // Determines if the error is about 'Not Found'
 export function isNotFoundError(error: any): boolean {
@@ -161,7 +75,9 @@ export async function getWrappedData(
 ): Promise<WrappedUploadResult> {
   try {
     const rawData = await bee.downloadData(eRef.toString(), options);
-    return rawData.toJSON() as WrappedUploadResult;
+    const wrappedResult = rawData.toJSON() as WrappedUploadResult;
+    asserWrappedUploadResult(wrappedResult);
+    return wrappedResult;
   } catch (error) {
     throw new FileInfoError(`Failed to get wrapped data: ${error}`);
   }
